requirements.txt
flask
sentence-transformers
faiss-cpu
gunicorn
numpy